pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "axiom-mcp"
version = "0.1.0"
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
description = "Model Context Protocol (MCP) implementation for connecting AI systems with external data sources"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.26.0",
    "mcp>=1.0.0,<2.0.0",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.5.3,<3.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=23.2.1",
    "aiohttp>=3.11.14",
    "cachetools>=5.3.2",
    "jinja2>=3.1.6",
    "aiosqlite>=0.21.0",
    # Dependencies for enhanced tools
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation>=0.42b0",
    "asyncio-throttle>=1.0.2",
    "redis>=5.0.1",  # For distributed caching
    "pyyaml>=6.0.1",  # For config loading
    "msgpack>=1.0.7",  # For efficient serialization
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0",
    "pytest-asyncio>=0.25.3",
]

db = [
    "asyncpg>=0.29.0",
    "aioredis>=2.0.1",
]

compression = [
    "lz4>=4.3.2",
    "zstandard>=0.22.0",
    "brotli>=1.1.0",
]

telemetry = [
    "opentelemetry-instrumentation-aiohttp-client>=0.42b0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-instrumentation-redis>=0.42b0",
]

validation = [
    "jsonschema>=4.21.1",
    "pydantic[email]>=2.5.3",
]

all = [
    "axiom-mcp[dev,db,compression,telemetry,validation]"
]

[project.scripts]
axiom-mcp = "axiom_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["axiom-mcp"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
select = ["E", "F", "B", "I"]
line-length = 88
target-version = "py312"
fix = true

[tool.ruff.isort]
known-first-party = ["axiom_mcp"]
combine-as-imports = true
force-wrap-aliases = true
profile = "black"

[tool.pytest.ini_options]
addopts = "--asyncio-mode=auto"
testpaths = ["tests"]
